complete CREATE, UPDATE, and DELETE

+++ key-value-store/HashTable.cpp
@@ -13,15 +13,16 @@ HashTable::~HashTable() {}
 /**
  * FUNCTION NAME: create
  *
- * DESCRIPTION: This function inserts they (key,value) pair into the local hash table
+ * DESCRIPTION: This function inserts they (key,value) pair into the local hash
+ * table
  *
  * RETURNS:
  * true on SUCCESS
  * false in FAILURE
  */
 bool HashTable::create(string key, string value) {
-	hashTable.emplace(key, value);
-	return true;
+  hashTable.emplace(key, value);
+  return true;
 }
 
 /**
@@ -34,66 +35,66 @@ bool HashTable::create(string key, string value) {
  * else it returns a NULL
  */
 string HashTable::read(string key) {
-	map<string, string>::iterator search;
+  map<string, string>::iterator search;
 
-	search = hashTable.find(key);
-	if ( search != hashTable.end() ) {
-		// Value found
-		return search->second;
-	}
-	else {
-		// Value not found
-		return "";
-	}
+  search = hashTable.find(key);
+  if (search != hashTable.end()) {
+    // Value found
+    return search->second;
+  } else {
+    // Value not found
+    return "";
+  }
 }
 
 /**
  * FUNCTION NAME: update
  *
- * DESCRIPTION: This function updates the given key with the updated value passed in
- * 				if the key is found
+ * DESCRIPTION: This function updates the given key with the updated value
+ * passed in if the key is found
  *
  * RETURNS:
  * true on SUCCESS
  * false on FAILURE
  */
 bool HashTable::update(string key, string newValue) {
-	map<string, string>::iterator update;
+  map<string, string>::iterator update;
 
-	if (read(key).empty()) {
-		// Key not found
-		return false;
-	}
-	// Key found
-	//update = hashTable.at(key) = newValue;
-	hashTable.at(key) = newValue;
-	// Update successful
-	return true;
+  if (read(key).empty()) {
+    // Key not found
+    return false;
+  }
+  // Key found
+  // update = hashTable.at(key) = newValue;
+  hashTable.at(key) = newValue;
+  // Update successful
+  return true;
 }
 
 /**
  * FUNCTION NAME: deleteKey
  *
- * DESCRIPTION: This function deletes the given key and the corresponding value if the key is found
+ * DESCRIPTION: This function deletes the given key and the corresponding value
+ * if the key is found
  *
  * RETURNS:
  * true on SUCCESS
  * false on FAILURE
  */
 bool HashTable::deleteKey(string key) {
-	uint eraseCount = 0;
+  uint16_t eraseCount = 0;
 
-	if (read(key).empty()) {
-		// Key not found
-		return false;
-	}
-	eraseCount = hashTable.erase(key);
-	if ( eraseCount < 1 ) {
-		// Could not erase
-		return false;
-	}
-	// Delete was successful
-	return true;
+  if (read(key).empty()) {
+    // Key not found
+    return false;
+  }
+  eraseCount = hashTable.erase(key);
+  if (eraseCount < 1) {
+    // Could not erase
+    return false;
+  }
+  // Delete was successful
+  return true;
 }
 
 /**
@@ -105,9 +106,7 @@ bool HashTable::deleteKey(string key) {
  * true if hash table is empty
  * false otherwise
  */
-bool HashTable::isEmpty() {
-	return hashTable.empty();
-}
+bool HashTable::isEmpty() { return hashTable.empty(); }
 
 /**
  * FUNCTION NAME: currentSize
@@ -118,7 +117,7 @@ bool HashTable::isEmpty() {
  * size of the table as unit
  */
 unsigned long HashTable::currentSize() {
-	return (unsigned  long)hashTable.size();
+  return (unsigned long)hashTable.size();
 }
 
 /**
@@ -126,9 +125,7 @@ unsigned long HashTable::currentSize() {
  *
  * DESCRIPTION: Clear all contents from the hash table
  */
-void HashTable::clear() {
-	hashTable.clear();
-}
+void HashTable::clear() { hashTable.clear(); }
 
 /**
  * FUNCTION NAME: count
@@ -139,6 +136,5 @@ void HashTable::clear() {
  * unsigned long count (Should be always 1)
  */
 unsigned long HashTable::count(string key) {
-	return (unsigned long) hashTable.count(key);
+  return (unsigned long)hashTable.count(key);
 }
-
